services/gsalegig-api/requirements.txt
blinker==1.8.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
Flask==3.0.3
googleapis-common-protos==1.59.1
grpcio==1.73.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-flask==0.55b1
opentelemetry-instrumentation-wsgi==0.55b1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
packaging==24.2
pi==0.1.2
protobuf==4.25.8
requests==2.32.3
setuptools==78.1.0
thrift==0.22.0
typing_extensions==4.14.0
urllib3==2.3.0
Werkzeug==3.0.4
wheel==0.46.1
wrapt==1.17.2
youtube-dl==2021.12.17
yt-dlp==2025.5.22
zipp==3.23.0
